# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ninja Kilat WMS",
    "description": "Warehouse management for Ninja Van's GrabMart Kilat fulfilment: inbound and putaway, guided picking, weekly stock opname, barcode and license-plate registration, and an isolated training mode.",
    "version": "0.1.0",
    "contact": {
      "name": "Ninja Van — Kilat fulfilment"
    }
  },
  "paths": {
...
    "/api/registry": {
      "get": {
        "tags": [
          "registry"
        ],
        "summary": "Every SKU with a pick face at this site, with its thresholds and stock",
        "description": "Every SKU with a pick face at this site, with its thresholds and stock.",
        "operationId": "list_registry_api_registry_get",
        "parameters": [
          {
            "name": "site_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Site Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "unconfigured_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Unconfigured Only"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "default": 200,
              "title": "Limit"
            }
          },
          {
            "name": "x-forwarded-email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Forwarded-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/registry/{sku_id}": {
      "put": {
        "tags": [
          "registry"
        ],
        "summary": "Set the three thresholds for one SKU's pick face",
        "description": "Set the three thresholds for one SKU's pick face.",
        "operationId": "set_thresholds_api_registry__sku_id__put",
        "parameters": [
          {
            "name": "sku_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sku Id"
            }
          },
          {
            "name": "x-forwarded-email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Forwarded-Email"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistryIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryRow"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/registry/bulk": {
      "post": {
        "tags": [
          "registry"
        ],
        "summary": "Apply one set of thresholds across many SKUs",
        "description": "Apply one set of thresholds across many SKUs.\n\n118 SKUs set one at a time will not happen, and a registry nobody fills in is\na registry that silently never fires a replenishment.",
        "operationId": "bulk_thresholds_api_registry_bulk_post",
        "parameters": [
          {
            "name": "x-forwarded-email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Forwarded-Email"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistryBulkIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/registry/suggest/{sku_id}": {
      "get": {
        "tags": [
          "registry"
        ],
        "summary": "Derive a starting point from the space model",
        "description": "Derive a starting point from the space model.\n\nA supervisor correcting a suggestion will fill this in; a supervisor facing\nthree empty boxes for 118 SKUs will not.",
        "operationId": "suggest_thresholds_api_registry_suggest__sku_id__get",
        "parameters": [
          {
            "name": "sku_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sku Id"
            }
          },
          {
            "name": "site_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Site Id"
            }
          },
          {
            "name": "x-forwarded-email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Forwarded-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrySuggestion"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Ok": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "Ok"
      },
...
      "RegistryBulkIn": {
        "properties": {
          "site_id": {
            "type": "integer",
            "title": "Site Id"
          },
          "sku_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Sku Ids"
          },
          "full_threshold": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Threshold"
          },
          "low_threshold": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Low Threshold"
          },
          "restock_point": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Restock Point"
          }
        },
        "type": "object",
        "required": [
          "site_id",
          "sku_ids"
        ],
        "title": "RegistryBulkIn"
      },
      "RegistryIn": {
        "properties": {
          "site_id": {
            "type": "integer",
            "title": "Site Id"
          },
          "full_threshold": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Threshold"
          },
          "low_threshold": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Low Threshold"
          },
          "restock_point": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Restock Point"
          }
        },
        "type": "object",
        "required": [
          "site_id"
        ],
        "title": "RegistryIn"
      },
      "RegistryList": {
        "properties": {
          "rows": {
            "items": {
              "$ref": "#/components/schemas/RegistryRow"
            },
            "type": "array",
            "title": "Rows"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "unconfigured": {
            "type": "integer",
            "title": "Unconfigured"
          }
        },
        "type": "object",
        "required": [
          "rows",
          "total",
          "unconfigured"
        ],
        "title": "RegistryList"
      },
      "RegistryRow": {
        "properties": {
          "sku_id": {
            "type": "integer",
            "title": "Sku Id"
          },
          "sku_name": {
            "type": "string",
            "title": "Sku Name"
          },
          "brand_sku_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand Sku Code"
          },
          "primary_location_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Location Id"
          },
          "primary_location_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Location Code"
          },
          "overflow_location_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overflow Location Id"
          },
          "overflow_location_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overflow Location Code"
          },
          "full_threshold": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Threshold"
          },
          "low_threshold": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Low Threshold"
          },
          "restock_point": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Restock Point"
          },
          "qty_primary": {
            "type": "integer",
            "title": "Qty Primary",
            "default": 0
          },
          "qty_overflow": {
            "type": "integer",
            "title": "Qty Overflow",
            "default": 0
          },
          "qty_total": {
            "type": "integer",
            "title": "Qty Total",
            "default": 0
          },
          "needs_replenishment": {
            "type": "boolean",
            "title": "Needs Replenishment",
            "default": false
          },
          "needs_restock": {
            "type": "boolean",
            "title": "Needs Restock",
            "default": false
          },
          "configured": {
            "type": "boolean",
            "title": "Configured",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "sku_id",
          "sku_name"
        ],
        "title": "RegistryRow"
      },
      "RegistrySuggestion": {
        "properties": {
          "sku_id": {
            "type": "integer",
            "title": "Sku Id"
          },
          "basket_size": {
            "type": "string",
            "title": "Basket Size"
          },
          "capacity_units": {
            "type": "integer",
            "title": "Capacity Units"
          },
          "full_threshold": {
            "type": "integer",
            "title": "Full Threshold"
          },
          "low_threshold": {
            "type": "integer",
            "title": "Low Threshold"
          },
          "restock_point": {
            "type": "integer",
            "title": "Restock Point"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "sku_id",
          "basket_size",
          "capacity_units",
          "full_threshold",
          "low_threshold",
          "restock_point",
          "reason"
        ],
        "title": "RegistrySuggestion"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}